requirements.txt
fastapi[standard]
numpy
tensorflow-cpu
pillow
prometheus-fastapi-instrumentator
mtcnn
opencv-python
uvicorn